updated comments for the climb arm io and moved it into the climb arm folder
