+++ Team-3749-2025/src/main/java/frc/robot/subsystems/arm/climb/ClimbArm.java
@@ -10,8 +10,7 @@ import edu.wpi.first.wpilibj.smartdashboard.MechanismRoot2d;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Robot;
-import frc.robot.subsystems.arm.ClimbArmIO;
-import frc.robot.subsystems.arm.ClimbArmIO.ArmData;
+import frc.robot.subsystems.arm.climb.ClimbArmIO.ArmData;
 import frc.robot.utils.ShuffleData;
 import frc.robot.utils.UtilityFunctions;
