+++ Team-3749-2025/src/main/java/frc/robot/subsystems/arm/climb/ClimbArmSparkMax.java
@@ -8,7 +8,6 @@ import com.revrobotics.spark.SparkMax;
 import com.revrobotics.spark.config.SparkBaseConfig.IdleMode;
 import com.revrobotics.spark.config.SparkMaxConfig;
 import edu.wpi.first.math.MathUtil;
-import frc.robot.subsystems.arm.ClimbArmIO;
 import frc.robot.subsystems.arm.ArmConstants;
 import frc.robot.utils.MiscConstants.SimConstants;
